make bow project (the whole modules) compatibles with linux
