+++ Sources/BowEffects/Foundation/FileManager+Common.swift
@@ -4,7 +4,7 @@ import Foundation
 @available(watchOS, unavailable)
 public extension FileManager {
     
-    // MARK: Locating System Directories
+    // MARK: - Locating System Directories
     
     /// IO suspended version of `FileManager.url(for:in:appropriateFor:create:)`. Refer to that method for further documentation.
     func urlIO(
@@ -17,7 +17,7 @@ public extension FileManager {
         }
     }
     
-    // MARK: Discovering Directory Contents
+    // MARK: - Discovering Directory Contents
     
     /// IO suspended version of `FileManager.contensOfDirectory(at:includingPropertiesForKeys:options:)`. Refer to that method for further documentation.
     func contentsOfDirectoryIO(
@@ -43,7 +43,7 @@ public extension FileManager {
         }
     }
     
-    // MARK: Creating and Deleting Items
+    // MARK: - Creating and Deleting Items
     
     /// IO suspended version of `FileManager.createDirectory(at:withIntermediateDirectories:attributes:)`. Refer to that method for further documentation.
     func createDirectoryIO(
@@ -89,17 +89,7 @@ public extension FileManager {
         }
     }
     
-    /// IO suspended version of `FileManager.trashItem(at:resultingItemURL:)`. Refer to that method for further documentation.
-    @available(iOS 11.0, *)
-    func trashItemIO(
-        at url: URL,
-        resultingItemURL outResultingURL: AutoreleasingUnsafeMutablePointer<NSURL?>?) -> IO<Error, ()> {
-        IO.invoke {
-            try self.trashItem(at: url, resultingItemURL: outResultingURL)
-        }
-    }
-    
-    // MARK: Replacing Items
+    // MARK: - Replacing Items
     
     /// IO suspended version of `FileManager.replaceItemAt(_:withItemAt:backupItemName:options:)`. Refer to that method for further documentation.
     func replaceItemIO(
@@ -150,57 +140,25 @@ public extension FileManager {
         }
     }
     
-    // MARK: Managing iCloud-Based Items
     
-    /// IO suspended version of `FileManager.setUbiquitous(_:itemAt:destinationURL:)`. Refer to that method for further documentation.
-    func setUbiquitousIO(
-        _ flag: Bool,
-        itemAt url: URL,
-        destinationURL: URL) -> IO<Error, ()> {
-        IO.invoke {
-            try self.setUbiquitous(flag, itemAt: url, destinationURL: destinationURL)
-        }
-    }
     
-    /// IO suspended version of `FileManager.startDownloadingUbiquitousItem(at:)`. Refer to that method for further documentation.
-    func startDownloadingUbiquitousItemIO(at url: URL) -> IO<Error, ()> {
-        IO.invoke {
-            try self.startDownloadingUbiquitousItem(at: url)
-        }
-    }
     
-    /// IO suspended version of `FileManager.evictUbiquitousItem(at:)`. Refer to that method for further documentation.
-    func evictUbiquitousItemIO(at url: URL) -> IO<Error, ()> {
-        IO.invoke {
-            try self.evictUbiquitousItem(at: url)
-        }
-    }
     
-    /// IO suspended version of `FileManager.url(forPublishingUbiquitousItemAt:expiration:)`. Refer to that method for further documentation.
-    func urlIO(
-        forPublishingUbiquitousItemAt url: URL,
-        expiration outDate: AutoreleasingUnsafeMutablePointer<NSDate?>?) -> IO<Error, URL> {
-        IO.invoke {
-            try self.url(forPublishingUbiquitousItemAt: url, expiration: outDate)
-        }
-    }
     
-    // MARK: Accessing File Provider Services
     
-    /// IO suspended version of `FileManager.getFileProviderServicesForItem(at:completionHandler:)`. Refer to that method for further documentation.
-    @available(OSX 10.13, *)
-    @available(iOS 11.0, *)
-    func getFileProviderServicesForItemIO(at url: URL) -> IO<Error, [NSFileProviderServiceName: NSFileProviderService]> {
-        IO.async { callback in
-            self.getFileProviderServicesForItem(at: url) { services, error in
-                if let services = services {
-                    callback(.right(services))
-                } else if let error = error {
-                    callback(.left(error))
-                }
-            }
-        }^
-    }
+    
+    
+    
+    
+    
+    
+    
+    
+    
+    
+    
+    
+    
     
     // MARK: Creating Symbolic and Hard Links
