--- Tests/LinuxMain.swift
@@ -1,6 +1,0 @@
-import XCTest
-@testable import bowTests
-
-XCTMain([
-    testCase(bowTests.allTests),
-])
